+++ typescript-client-sdk/src/sdk/sdk.ts
@@ -47,8 +47,8 @@ export class SDK {
     public _securityClient: AxiosInstance;
     public _serverURL: string;
     private _language = "typescript";
-    private _sdkVersion = "0.3.0";
-    private _genVersion = "2.32.7";
+    private _sdkVersion = "0.4.0";
+    private _genVersion = "2.34.7";
     private _globals: any;
 
     constructor(props?: SDKProps) {