+++ RELEASES.md
@@ -189,3 +189,13 @@ Based on:
 - Speakeasy CLI 1.39.0 (2.32.7) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [NPM v0.3.0] https://www.npmjs.com/package/brease-sdk/v/0.3.0 - typescript-client-sdk
+
+## 2023-06-02 00:47:45
+### Changes
+Based on:
+- OpenAPI Doc 0.1.0 https://prod.speakeasyapi.dev/v1/apis/brease/version/0.1/schema/download
+- Speakeasy CLI 1.40.3 (2.34.7) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [Go v0.7.0] https://github.com/speakeasy-sdks/brease/releases/tag/go-client-sdk/v0.7.0 - go-client-sdk
+- [PyPI v1.3.0] https://pypi.org/project/brease-sdk/1.3.0 - python-client-sdk
+- [NPM v0.4.0] https://www.npmjs.com/package/brease-sdk/v/0.4.0 - typescript-client-sdk